* changed type for sizeof_type in create routines to size_t

+++ src/queue/struct/mulle--structqueue.c
@@ -40,7 +40,7 @@ int   __mulle__structqueueenumerator_next( struct mulle__structqueueenumerator *
 
 
 struct mulle__structqueue   *
-   mulle__structqueue_create( unsigned int sizeof_struct,
+   mulle__structqueue_create( size_t sizeof_struct,
                               unsigned int alignof_struct,
                               unsigned short bucket_size,
                               unsigned short spare_allowance,
@@ -103,7 +103,7 @@ void
 
 
 void   _mulle__structqueue_grow( struct mulle__structqueue *queue,
-                                  struct mulle_allocator *allocator)
+                                 struct mulle_allocator *allocator)
 {
    struct mulle__structqueuebucket   *p;
    struct mulle__structqueuebucket   *q;
